Phase 5.5: two clippy findings fixed, and the results doc
Both findings were fixed rather than suppressed, following Phase 4's
precedent for the same lint. The nine-argument `compare` in the
differential became a `Request` struct and a method on the two-sided
fixture, which is better than a suppression for a reason specific to this
test: a differential whose two arms are fed by two separate argument
lists can compare two different requests and call it agreement, and one
struct consumed by both arms removes that failure mode.

Clippy is clean from a genuine cold build -- 113 units checked from an
empty target, all five workspace crates named in the output, exit 0, zero
warnings, verified by reading the captured log rather than by the exit
code. Zero suppressions added anywhere under `crates/`, checked by
running the diff.

493 tests passing, 0 failed, 0 ignored, 0 skipped. `src/`,
`docs/gpu-native-architecture.md` and the root `Cargo.toml` are all
untouched, confirmed by `git diff` rather than asserted.

`docs/phase-5.5-results.md` has the full account, including the two
places the phase turned out not to be a pure transcription, the honest
read on what "it renders" does and does not mean yet (correct glyph
pixels on the GPU; still nothing that binds and draws them), and the
legacy sub-pixel quirk this port reproduces on purpose.

+++ crates/wgpui-wgpu/src/render/atlas_upload.rs
@@ -280,11 +280,9 @@ mod tests {
         // other rather than against a comment.
         for kind in [AtlasKind::Monochrome, AtlasKind::Polychrome] {
             assert_eq!(
-                u32::from(
-                    texture_format(kind)
-                        .block_copy_size(None)
-                        .expect("an uncompressed colour format has a block size")
-                ),
+                texture_format(kind)
+                    .block_copy_size(None)
+                    .expect("an uncompressed colour format has a block size"),
                 kind.bytes_per_pixel()
             );
         }
